% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cauchy.2.R
\name{cauchy.2}
\alias{cauchy.2}
\title{Generate random samples from Cauchy distribution by MH sampler}
\usage{
cauchy.2(n, theta = 1, eta = 0)
}
\arguments{
\item{n}{number of observations.}

\item{theta}{location parameter}

\item{eta}{scale parameter}
}
\value{
a random sample of size \code{n}
}
\description{
Generate random samples from Cauchy distribution by MH sampler
}
\examples{
\dontrun{
cauchy.2(100)
}
}
