% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rayleigh.1.R
\name{Rayleigh.1}
\alias{Rayleigh.1}
\title{Generate samples from Rayleigh distribution}
\usage{
Rayleigh.1(x, sigma)
}
\arguments{
\item{x}{vector of quantiles}

\item{sigma}{the parameter of Rayleigh density}
}
\value{
a value
}
\description{
Generate samples from Rayleigh distribution
}
\examples{
\dontrun{
Rayleigh.1(10,1)
}
}
